converted certainties to percentages, updated relationship forms and views:

+++ db/migrate/20140612231611_create_user_rel_contribs.rb
@@ -3,7 +3,7 @@ class CreateUserRelContribs < ActiveRecord::Migration
     create_table :user_rel_contribs do |t|
       t.integer :relationship_id
       t.integer :created_by
-      t.decimal :confidence
+      t.integer :confidence
       t.text :annotation
       t.text :bibliography
       t.integer :relationship_type
